% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greek_myth_color_package.R
\name{GreekMyth_palette}
\alias{GreekMyth_palette}
\title{Greek Myth Palette Generator}
\usage{
GreekMyth_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of Palette.}

\item{n}{Number of desired colors. If number of requested colors is beyond the scope of the palette,
colors are automatically interpolated. If n is not provided, the length of the palette is used.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want to automatically
interpolate between colors.}
}
\value{
A vector of colors.
}
\description{
Color palettes inspired by paintings depicting stories of Greek mythology.
}
\examples{
GreekMyth_palette("Birth_Venus")

GreekMyth_palette("Birth_Venus", n=3)

library(ggplot2)
ggplot(data=iris, aes(x=Species, y=Petal.Length, fill=Species)) +
geom_violin() +
scale_fill_manual(values=GreekMyth_palette("Birth_Venus", 3))

ggplot(data=iris, aes(x=Sepal.Length, y=Sepal.Width, color=Species)) +
geom_point(size=2) +
scale_color_manual(values=GreekMyth_palette("Birth_Venus", 3))

}
\keyword{colors}
