% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greek_myth_color_package.R
\name{GreekMyth_story}
\alias{GreekMyth_story}
\title{Get Long Text Description}
\usage{
GreekMyth_story(name)
}
\arguments{
\item{name}{Name of the variable}
}
\value{
A long text description
}
\description{
Get Long Text Description
}
\examples{
GreekMyth_story("Birth_Venus")
}
