% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greek_myth_color_package.R
\docType{data}
\name{myth_palettes}
\alias{myth_palettes}
\title{Complete list of palettes
Use names(myth_palettes) to return all possible palette names.
Use \code{\link{myth_palettes}} to construct palettes of desired length.}
\format{
An object of class \code{list} of length 17.
}
\usage{
myth_palettes
}
\description{
Complete list of palettes
Use names(myth_palettes) to return all possible palette names.
Use \code{\link{myth_palettes}} to construct palettes of desired length.
}
\keyword{datasets}
